Extracted microservice status into separate  comp
