+++ Source/SelfService/Web/microservice/dataTable.tsx
@@ -6,27 +6,10 @@ import { useHistory } from 'react-router-dom';
 
 import { getPodStatus, MicroserviceInfo } from '../api/api';
 import { HttpResponseApplication } from '../api/application';
-import { MicroserviceStatus, getMicroserviceState } from './microserviceStatus';
+import { statusCell, customStatusFieldSort } from './microserviceStatus';
 
 import { DataGridPro, GridColDef, GridValueGetterParams, GridRenderCellParams } from '@mui/x-data-grid-pro';
-
-import { Box, Paper, Tooltip, Theme, Typography } from '@mui/material';
-import { CheckCircleRounded, ErrorRounded, WarningRounded, QuestionMark } from '@mui/icons-material';
-
-const styles = {
-    status: {
-        display: 'flex',
-        justifyContent: 'center'
-    },
-    statusTitle: {
-        fontWeight: 500,
-        fontSize: '0.75rem',
-        lineHeight: '1.375rem',
-        letterSpacing: '0.06rem',
-        textTransform: 'uppercase',
-        ml: 1.25
-    }
-};
+import { Box, Paper, Tooltip } from '@mui/material';
 
 const sortByRuntimeVersion = (params: GridValueGetterParams) => {
     const runtimeVersion = params.row.edit?.extra?.runtimeImage?.replace(/dolittle\/runtime:/gi, '');
@@ -50,67 +33,6 @@ const publicUrlCell = (params: GridRenderCellParams) => {
     );
 };
 
-// enum MicroserviceStatus {
-//     Running = 0,
-//     Pending = 1,
-//     Failing = 2,
-//     Unknown = 3,
-// }
-
-// const getMicroserviceState = (phase?: string): MicroserviceStatus => {
-//     const checkStatus = phase?.toLowerCase?.();
-
-//     if (typeof checkStatus !== 'string') {
-//         return MicroserviceStatus.Unknown;
-//     } else if (checkStatus.includes('running')) {
-//         return MicroserviceStatus.Running;
-//     } else if (checkStatus.includes('pending')) {
-//         return MicroserviceStatus.Pending;
-//     } else if (checkStatus.includes('failed')) {
-//         return MicroserviceStatus.Failing;
-//     }
-
-//     return MicroserviceStatus.Unknown;
-// };
-
-const statusCell = (params: GridRenderCellParams) => {
-    let color = (theme: Theme) => theme.palette.text.primary;
-    let icon = <QuestionMark sx={{ color }} />;
-    let status = 'N/A';
-
-    switch (getMicroserviceState(params.row.status[0]?.phase)) {
-        case MicroserviceStatus.Running:
-            icon = <CheckCircleRounded />;
-            status = 'Running';
-            break;
-        case MicroserviceStatus.Pending:
-            color = (theme: Theme) => theme.palette.warning.main;
-            icon = <WarningRounded sx={{ color }} />;
-            status = 'Pending';
-            break;
-        case MicroserviceStatus.Failing:
-            color = (theme: Theme) => theme.palette.error.main;
-            icon = <ErrorRounded sx={{ color }} />;
-            status = 'Failed';
-            break;
-        case MicroserviceStatus.Unknown:
-            return 'N/A';
-    }
-
-    return (
-        <Box sx={styles.status}>
-            {icon}
-            <Typography sx={{ ...styles.statusTitle, color }}>{status}</Typography>
-        </Box>
-    );
-};
-
-const customStatusFieldSort = (_, __, left, right) => {
-    const leftStatus = getMicroserviceState(left.value[0]?.phase);
-    const rightStatus = getMicroserviceState(right.value[0]?.phase);
-    return leftStatus - rightStatus;
-};
-
 export type MicroserviceObject = {
     id: string
     name: string
@@ -122,7 +44,7 @@ export type MicroserviceObject = {
             isPublic: boolean
         }
     }
-    status?: []
+    state?: string
 };
 
 type DataTableProps = {
@@ -146,7 +68,7 @@ export const DataTable = ({ application, environment, microservices }: DataTable
 
             return {
                 ...microservice,
-                status
+                state: status[0]?.phase
             } as MicroserviceObject;
         }))
             .then(setRows);
@@ -196,7 +118,7 @@ export const DataTable = ({ application, environment, microservices }: DataTable
             sortComparator: customUrlFieldSort
         },
         {
-            field: 'status',
+            field: 'state',
             headerName: 'Status',
             minWidth: 200,
             flex: 1,
